test: Match rest of style
